requirements.txt
flask
flask-cors
requests
gunicorn
openai
python-dotenv
